fin ajout de la bdd

+++ bdd/setup.sh
@@ -1,6 +1,10 @@
-#!/bin/bash
+#!/bin/sh
 
 echo "Setting up the environment for the BDD tests"
 
-echo "Create environnement file"
-cp .env.example .env
+if [ -f ".env" ]; then
+    echo "Create environnement file"
+    cp .env.example .env
+elif 
+    echo "Environnement file already exists"
+fi
